add `ui/time-current` tests
